root: added k8s binary authorization notes
binary authorization seems to be a new security measure (not bad
in fact). Despite its complexity, it is a necessary evil. Hence,
let's go ahead and add the notes in and improve it in the future.

This patch adds k8s binary authorization notes into many parts of
root repository.

Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>