cases/qwiklab-jooli-data-junior.sh: fixed AI bug
Appearently, the assessment tool reads metadata as well so it has
to be 'application/json' and not 'application/octet-stream'. Hence,
I need to fix the bug.

Also, the naming of the file is wrong for the speech API. It needs
patching as well.

This patch fix AI bug in cases/qwiklab-jooli-data-junior.sh in
order to pass the challenge.

Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>