root: added deployment-manager notes
Since cloud architect employs the use of jinja deployment manager
for VM fleet and it's rare that it will be used (since there is
k8s to do the job), it's best I add a sample notes here so that
there is something to reference from.

This patch adds deployment-manager notes into root directory.

Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>